add apis in messenge board
